split the grid and space classes
—
